added test to see if any files need renaming
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@3315 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ scripts/convert-help-to-standard.sh
@@ -10,8 +10,10 @@
 # renamed by their respective maintainers.
 
 for helpfile in help-*.pd ; do
-   mv "$helpfile" "`echo $helpfile | sed 's/^help\-\(.*\)\.pd$/\1-help.pd/'`"
-#	echo "SWAP $helpfile `echo $helpfile | sed 's/^help\-\(.*\)\.pd$/\1-help.pd/'`"
+	 if [ -e $helpfile ]; then
+		  mv "$helpfile" "`echo $helpfile | sed 's/^help\-\(.*\)\.pd$/\1-help.pd/'`"
+		  echo "SWAP $helpfile FOR `echo $helpfile | sed 's/^help\-\(.*\)\.pd$/\1-help.pd/'`"
+	 fi
 done
 
 # this isn't used yet
